+++ services/cdnjs/cdnjs.service.js
@@ -6,12 +6,17 @@ const { NotFound } = require('../errors')
 const { addv: versionText } = require('../../lib/text-formatters')
 const { version: versionColor } = require('../../lib/color-formatters')
 
+const cdnjsSchema = Joi.object({
+  // optional due to non-standard 'not found' condition
+  version: Joi.string(),
+}).required()
+
 module.exports = class Cdnjs extends BaseJsonService {
   async handle({ library }) {
     const url = `https://api.cdnjs.com/libraries/${library}?fields=version`
     const json = await this._requestJson({
       url,
-      schema: Joi.any(),
+      schema: cdnjsSchema,
     })
 
     if (Object.keys(json).length === 0) {
@@ -19,11 +24,10 @@ module.exports = class Cdnjs extends BaseJsonService {
          status code = 200, body = {} */
       throw new NotFound()
     }
-    const version = json.version || 0
 
     return {
-      message: versionText(version),
-      color: versionColor(version),
+      message: versionText(json.version),
+      color: versionColor(json.version),
     }
   }
